add updateStatusContact method
