+++ routes/api/contacts.js
@@ -5,6 +5,7 @@ import {
   getListContacts,
   postNewContact,
   putContact,
+  updateStatusContact,
 } from "../../controllers/controllers.js";
 
 export const contactsRouter = express.Router();
@@ -18,5 +19,7 @@ contactsRouter.post("/", postNewContact);
 contactsRouter.delete("/:contactId", deleteContact);
 
 contactsRouter.put("/:contactId", putContact);
+
+contactsRouter.patch("/:contactId/favorite", updateStatusContact);
 
 // simakov hJ56BT8zt7chUBeV
